Move Finished.js
